update utils imports
