+++ pyxem/signals/indexation_results.py
@@ -21,10 +21,9 @@ import hyperspy.api as hs
 from hyperspy.signal import BaseSignal
 from warnings import warn
 
-from pyxem.signals import push_metadata_through
-from pyxem.utils.sim_utils import peaks_from_best_template
-from pyxem.utils.sim_utils import peaks_from_best_vector_match
-from pyxem.utils.sim_utils import transfer_navigation_axes
+from pyxem.signals import push_metadata_through, transfer_navigation_axes
+from pyxem.utils.indexation_utils import peaks_from_best_template
+from pyxem.utils.indexation_utils import peaks_from_best_vector_match
 from pyxem.utils.indexation_utils import crystal_from_template_matching
 from pyxem.utils.indexation_utils import crystal_from_vector_matching
 from pyxem.utils.plot import generate_marker_inputs_from_peaks
