+++ pyxem/tests/test_physical/test_orientation_mapping_phys.py
@@ -24,16 +24,16 @@ import diffpy.structure
 
 from transforms3d.euler import euler2mat
 
-from pyxem.generators.indexation_generator import IndexationGenerator
-from pyxem.generators.indexation_generator import VectorIndexationGenerator
 from diffsims.generators.library_generator import VectorLibraryGenerator
 from diffsims.libraries.structure_library import StructureLibrary
+from diffsims.utils.sim_utils import get_kinematical_intensities
+
+from pyxem.generators.indexation_generator import IndexationGenerator
+from pyxem.generators.indexation_generator import VectorIndexationGenerator
 from pyxem.signals.electron_diffraction import ElectronDiffraction
 from pyxem.signals.diffraction_vectors import DiffractionVectors
-from pyxem.utils.sim_utils import (peaks_from_best_template,
-                                   peaks_from_best_vector_match,
-                                   get_kinematical_intensities)
-
+from pyxem.utils.indexation_utils import peaks_from_best_template
+from pyxem.utils.indexation_utils import peaks_from_best_vector_match
 from pyxem.utils.sim_utils import sim_as_signal
 
 """
