+++ pyxem/tests/test_utils/test_sim_utils.py
@@ -26,8 +26,8 @@ from diffsims.utils.sim_utils import get_electron_wavelength, \
     get_vectorized_list_for_atomic_scattering_factors, get_points_in_sphere, \
     simulate_kinematic_scattering, is_lattice_hexagonal, uvtw_to_uvw, \
     rotation_list_stereographic
-from pyxem.utils.sim_utils import transfer_navigation_axes, \
-    peaks_from_best_template
+from pyxem.signals import transfer_navigation_axes
+from pyxem.signals import peaks_from_best_template
 
 
 def create_lattice_structure(a, b, c, alpha, beta, gamma):
